fastapi/film/src/pyproject.toml
[tool.poetry]
name = "fastapi films"
version = "0.1.0"
description = "async api sprint 1"
authors = ["Sergey Bilibin", "Alexandr Bondarenko", "Tatyana Ovsyankina"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
redis="==4.4.2"
elasticsearch="==7.17"
aiohttp="==3.8.4"
fastapi=">=0.68.0"
orjson="==3.4.7"
pydantic="==1.9.0"
uvicorn="==0.12.2"
uvloop="==0.17.0"
python-dotenv="==1.0.0"
fastapi-pagination="==0.9.0"
gunicorn="==20.1.0"
httptools="==0.5.0"
httpx="==0.23.3"

[tool.poetry.dev-dependencies]
pytest = "==7.2.1"
isort = "==5.12.0"
flake8 = "==6.0.0"
mypy = "==1.0.0"
pre-commit = "==3.0.4"
Flake8-pyproject = "==1.2.2"
pyproject-autoflake = "==1.0.0"

[tool.autoflake]
in-place=true
expand-star-imports=true
remove-duplicate-keys=true
remove-all-unused-imports=true

[tool.black]
line-length = 120
skip-string-normalization=false

[tool.mypy]
disallow-untyped-defs=true
disallow-incomplete-defs=true
disallow-untyped-calls=true      


[tool.isort]
multi_line_output=3
include_trailing_comma=true
force_grid_wrap=0
use_parentheses=true
line_length=120

[tool.flake8]
ignore = ['E203', 'E266', 'E501', 'W503']
max-line-length = 120
max-complexity = 18
select = ['B','C','E','F','W','T4']
exclude= [    
    '.env.example',
    '.env',
    '.env.dev'
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
